+++ client/iOS/Helper/UIImage+HockeyAdditions.h
@@ -29,7 +29,7 @@
 
 - (UIImage *)bw_roundedCornerImage:(NSInteger)cornerSize borderSize:(NSInteger)borderSize;
 - (UIImage *)bw_imageToFitSize:(CGSize)fitSize honorScaleFactor:(BOOL)honorScaleFactor;
-- (UIImage *)bw_reflectedImageWithHeight:(NSUInteger)height fromAlpha:(float)fromAlpha toAlpha:(float)toAlpha;
+- (UIImage *)bw_reflectedImageWithHeight:(CGFloat)height fromAlpha:(CGFloat)fromAlpha toAlpha:(CGFloat)toAlpha;
 
 - (id)bw_initWithContentsOfResolutionIndependentFile:(NSString *)path NS_RETURNS_RETAINED;
 + (UIImage*)bw_imageWithContentsOfResolutionIndependentFile:(NSString *)path;
